test: #5531 EPS 회귀를 integration suite로 이동
